+++ test/libsolidity/syntaxTests/bytecodeReferences/circular_reference_complex.sol
@@ -1,0 +1,7 @@
+contract D {}
+contract C is D {}
+contract E is D
+{
+	function foo() public { new C(); }
+}
+// ----