+++ test/libsolidity/syntaxTests/bytecodeReferences/circular_reference_function_parameters.sol
@@ -1,0 +1,3 @@
+contract C { function foo(D _d) public { _d.foo(this); } }
+contract D { function foo(C _c) public { _c.foo(this); } }
+// ----