+++ test/libsolidity/syntaxTests/bytecodeReferences/circular_reference_new_in_ctor.sol
@@ -1,0 +1,5 @@
+contract C {
+	constructor() { new C(); }
+}
+// ----
+// TypeError 7813: (30-35): Circular reference to contract bytecode either via "new" or "type(...).creationCode" / "type(...).runtimeCode".